requirements.txt
fastapi==0.117.1
uvicorn[standard]==0.36.0
python-multipart==0.0.20
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-dotenv==1.1.1
pydantic==2.11.9
pydantic-settings==2.6.1

# ORM 추천 패키지들 (설치하지 않음, 참고용)
# SQLAlchemy==2.0.36
# alembic==1.14.0
# asyncpg==0.30.0  # PostgreSQL
# psycopg2-binary==2.9.10  # PostgreSQL
# aiomysql==0.2.0  # MySQL
# motor==3.6.0  # MongoDB

# 개발 및 테스트
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1
